alumni/requirements.txt
Django==1.9.6
djangorestframework==3.3.3
Pillow==3.2.0
requests==2.9.1
South==1.0.2
Unirest==1.1.7
WebError==0.13.1
WebHelpers==1.3
django-smart-selects==1.2.2